Adaptions to last changes in Cachebox Core

+++ CB_Server/src/cb_server/IconServlet.java
@@ -1,18 +1,12 @@
 package cb_server;
 
 import java.awt.BasicStroke;
-import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.Image;
-import java.awt.Toolkit;
 import java.awt.image.BufferedImage;
 import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
 import java.net.URL;
 
 import javax.imageio.ImageIO;
@@ -31,8 +25,6 @@ import org.mapsforge.core.graphics.Style;
 import org.mapsforge.core.graphics.TileBitmap;
 import org.mapsforge.map.awt.AwtGraphicFactory;
 
-import CB_Core.Enums.CacheTypes;
-
 public class IconServlet extends HttpServlet {
 	private static final long serialVersionUID = 1205779103262021876L;
 
@@ -40,8 +32,7 @@ public class IconServlet extends HttpServlet {
 
 	}
 
-	public void doGet___(HttpServletRequest req, HttpServletResponse resp)
-			throws IOException {
+	public void doGet___(HttpServletRequest req, HttpServletResponse resp) throws IOException {
 
 		String fileName = "/icons/32-0.png";
 		Image img = getImage(fileName);
@@ -54,8 +45,7 @@ public class IconServlet extends HttpServlet {
 		Graphics2D graphics = (Graphics2D) image.getGraphics();
 		graphics.setColor(new java.awt.Color(1.0f, 1.0f, 1.0f, 0.75f));
 		graphics.fillRoundRect(0, 0, 32, 32, 5, 5);
-		graphics.drawImage(img, 0, 0, image.getWidth(), image.getHeight(), 0,
-				0, img.getWidth(null), img.getHeight(null), null);
+		graphics.drawImage(img, 0, 0, image.getWidth(), image.getHeight(), 0, 0, img.getWidth(null), img.getHeight(null), null);
 		graphics.dispose();
 
 		if (ImageIO.write(image, "png", resp.getOutputStream()))
@@ -64,8 +54,7 @@ public class IconServlet extends HttpServlet {
 	}
 
 	@Override
-	protected void doGet(HttpServletRequest request,
-			HttpServletResponse response) throws ServletException, IOException {
+	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		synchronized (this) {
 			ByteArrayOutputStream imgOutputStream = new ByteArrayOutputStream();
 			byte[] captchaBytes = null; // imageBytes
@@ -135,8 +124,7 @@ public class IconServlet extends HttpServlet {
 				String postfix = "";
 				if (solved)
 					postfix = "S";
-				String fileName = "/icons/" + prefix + cacheType + postfix
-						+ ".png";
+				String fileName = "/icons/" + prefix + cacheType + postfix + ".png";
 				if (found) {
 					fileName = "/icons/" + prefix + "Found.png";
 				}
@@ -152,35 +140,29 @@ public class IconServlet extends HttpServlet {
 				BufferedImage image = null;
 
 				if (diffTerr) {
-					String id = String.valueOf((int) (difficulty / 2));
+					String id = String.valueOf(difficulty / 2);
 					if (difficulty % 2 > 0) {
 						id += "-5";
 					}
-					String id2 = String.valueOf((int) (terrain / 2));
+					String id2 = String.valueOf(terrain / 2);
 					if (terrain % 2 > 0) {
 						id2 += "-5";
 					}
 					img = getImage("/icons/stars" + id + "small.png");
 
 					int starHeight = img.getWidth(null);
-					int starWidth = Math
-							.round((float) (img.getHeight(null) * size)
-									/ starHeight);
+					int starWidth = Math.round((float) (img.getHeight(null) * size) / starHeight);
 					background = size + 2 * starWidth;
-					image = new BufferedImage(starWidth, size,
-							BufferedImage.TYPE_INT_ARGB);
+					image = new BufferedImage(starWidth, size, BufferedImage.TYPE_INT_ARGB);
 					Graphics2D graphics = (Graphics2D) image.getGraphics();
 					graphics.setColor(new java.awt.Color(1.0f, 1.0f, 1.0f, 0.5f));
-					graphics.fillRoundRect(0, 0, starWidth + 2, starHeight,
-							starWidth / 2, starWidth / 2);
+					graphics.fillRoundRect(0, 0, starWidth + 2, starHeight, starWidth / 2, starWidth / 2);
 					// graphics.fillRoundRect(background - 2 - starWidth, 0,
 					// background, starHeight, starWidth / 2, starWidth / 2);
 					int dx = -size;
 					int dy = 1;
 					graphics.rotate(-Math.PI / 2);
-					graphics.drawImage(img, dx, dy, dx + size - 1, dy
-							+ starWidth, 0, 0, img.getWidth(null),
-							img.getHeight(null), null);
+					graphics.drawImage(img, dx, dy, dx + size - 1, dy + starWidth, 0, 0, img.getWidth(null), img.getHeight(null), null);
 					dy += size + starWidth - 2;
 					// graphics.drawImage(img2, dx, dy, dx + size - 1, dy +
 					// starWidth, 0, 0, img.getWidth(null), img.getHeight(null),
@@ -199,19 +181,14 @@ public class IconServlet extends HttpServlet {
 						img = getImage(fileName);
 					}
 
-					image = new BufferedImage(background, background,
-							BufferedImage.TYPE_INT_ARGB);
+					image = new BufferedImage(background, background, BufferedImage.TYPE_INT_ARGB);
 					Graphics2D graphics = (Graphics2D) image.getGraphics();
 
-					graphics.drawImage(img, 0, 0, image.getWidth(),
-							image.getHeight(), 0, 0, img.getWidth(null),
-							img.getHeight(null), null);
+					graphics.drawImage(img, 0, 0, image.getWidth(), image.getHeight(), 0, 0, img.getWidth(null), img.getHeight(null), null);
 					if (img2 != null) {
 						int dx = (background - size) / 2;
 						int dy = (background - size) / 2;
-						graphics.drawImage(img2, dx, dy, dx + size, dx + size,
-								0, 0, img2.getWidth(null),
-								img2.getHeight(null), null);
+						graphics.drawImage(img2, dx, dy, dx + size, dx + size, 0, 0, img2.getWidth(null), img2.getHeight(null), null);
 					}
 					if (deactivated || archived) {
 						// Roter durchstreichen
@@ -221,11 +198,9 @@ public class IconServlet extends HttpServlet {
 							rand = 2;
 							width = 2;
 						}
-						graphics.setStroke(new BasicStroke(width,
-								BasicStroke.CAP_ROUND, BasicStroke.CAP_ROUND));
+						graphics.setStroke(new BasicStroke(width, BasicStroke.CAP_ROUND, BasicStroke.CAP_ROUND));
 						graphics.setColor(java.awt.Color.red);
-						graphics.drawLine(rand, rand, background - rand - 1,
-								background - rand - 1);
+						graphics.drawLine(rand, rand, background - rand - 1, background - rand - 1);
 					}
 					graphics.dispose();
 				}
@@ -260,8 +235,7 @@ public class IconServlet extends HttpServlet {
 		}
 	}
 
-	protected void doGet_(HttpServletRequest request,
-			HttpServletResponse response) throws ServletException, IOException {
+	protected void doGet_(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		String sQuery = request.getPathInfo();
 		String[] query = sQuery.split("/");
 		int size = Integer.parseInt(query[1]);
@@ -302,8 +276,7 @@ public class IconServlet extends HttpServlet {
 		ResourceBitmap bmp2 = null;
 		if (backgroundSize > 0) {
 			if (selected) {
-				is = getClass().getResourceAsStream(
-						"/icons/shaddowrect-selected.png");
+				is = getClass().getResourceAsStream("/icons/shaddowrect-selected.png");
 			} else {
 				is = getClass().getResourceAsStream("/icons/shaddowrect.png");
 			}
@@ -316,20 +289,16 @@ public class IconServlet extends HttpServlet {
 			backgroundSize = size;
 		}
 
-		TileBitmap bitmap = AwtGraphicFactory.INSTANCE.createTileBitmap(
-				backgroundSize, true);
+		TileBitmap bitmap = AwtGraphicFactory.INSTANCE.createTileBitmap(backgroundSize, true);
 		Canvas canvas = AwtGraphicFactory.INSTANCE.createCanvas();
 		canvas.setBitmap(bitmap);
 		Matrix matrix = AwtGraphicFactory.INSTANCE.createMatrix();
-		matrix.scale((float) backgroundSize / (float) bmp.getWidth(),
-				(float) backgroundSize / (float) bmp.getHeight());
+		matrix.scale((float) backgroundSize / (float) bmp.getWidth(), (float) backgroundSize / (float) bmp.getHeight());
 		canvas.drawBitmap(bmp, matrix);
 		if (bmp2 != null) {
 			matrix = AwtGraphicFactory.INSTANCE.createMatrix();
-			matrix.translate((backgroundSize - size) / 2,
-					(backgroundSize - size) / 2);
-			matrix.scale((float) size / (float) bmp2.getWidth(), (float) size
-					/ (float) bmp2.getHeight());
+			matrix.translate((backgroundSize - size) / 2, (backgroundSize - size) / 2);
+			matrix.scale((float) size / (float) bmp2.getWidth(), (float) size / (float) bmp2.getHeight());
 			canvas.drawBitmap(bmp2, matrix);
 		}
 		if (deactivated) {
